+++ macros/file_formats/stage_file_formats.sql
@@ -1,6 +1,6 @@
 {% macro stage_file_formats(enabled_targets) %}
-    {% if target.name in enabled_targets %}
-        {% if flags.WHICH == 'run' or flags.WHICH == 'run-operation' %}
+    {% if flags.WHICH == 'run' or flags.WHICH == 'run-operation' %}
+        {% if target.name in enabled_targets %}
             {% set items_to_stage = [] %}
 
             {% set nodes = graph.nodes.values() if graph.nodes else [] %}
@@ -16,10 +16,11 @@
             {% if items_to_stage|length > 0 %}
                 {% do dbt_dataengineers_materializations.stage_file_format_plans(items_to_stage) %}
             {% endif %}
+        {% else %}
+            {% do log('file formats to create: Not enabled', info = true) %}
         {% endif %}
-    {% else %}
-        {% do log('file formats to create: Not enabled', info = true) %}
     {% endif %}
+
 {% endmacro %}
 
 
@@ -27,10 +28,10 @@
     {% for node in items_to_stage %}
         {% set loop_label = loop.index ~ ' of ' ~ loop.length %}
         {% do log(loop_label ~ ' START file format creation ' ~ node.schema ~ '.' ~ node.name, info = true) -%}
-        
+
         {% set run_queue = dbt_dataengineers_materializations.get_file_format_build_plan(node) %}
         {% do log(loop_label ~ ' SKIP file format ' ~ node.schema ~ '.' ~ node.name, info = true) if run_queue == [] %}
-        
+
         {% set width = flags.PRINTER_WIDTH %}
         {% for cmd in run_queue %}
             {# do log(loop_label ~ ' ' ~ cmd, info = true) #}