+++ macros/source_tables/snowflake/snowflake__get_source_build_plan.sql
@@ -1,4 +1,4 @@
-{% macro snowflake__get_source_build_plan(source_node, is_first_run) %}
+{% macro snowflake__get_source_build_plan(source_node, is_first_run, auto_maintained) %}
     {% set build_plan = [] %}
 
     {# Setup our variables which are re-usable #}
@@ -23,6 +23,7 @@
         {%- set stream_name = dbt_dataengineers_materializations.snowflake_get_stream_name(identifier) -%}
         {%- set stream_relation = api.Relation.create(schema=schema, identifier=stream_name) -%}
 
+        {% do build_plan.append(dbt_dataengineers_materializations.snowflake_create_schema(target_relation)) %}
         {# determine if we need to replace a view with this table #}
         {% if current_relation_exists_as_view %}
             {% do build_plan.append(dbt_dataengineers_materializations.snowflake_drop_view(current_relation)) %}
@@ -30,7 +31,7 @@
         {% endif %}
 
         {# determine backups and mirgation and comparison tables accoridngly #}
-        {% if current_relation_exists_as_table %}
+        {% if current_relation_exists_as_table and auto_maintained %}
             {% if source_node.external.retain_previous_version_flg %}
                 {%- set backup_suffix_dt = py_current_timestring() -%}
                 {%- set backup_table_suffix = config.get('backup_table_suffix', default='_DBT_BACKUP_') -%}
@@ -47,11 +48,11 @@
                 {% do build_plan.append(dbt_dataengineers_materializations.snowflake_drop_table(current_relation)) %}
             {% endif %}
             {% do build_plan.append(dbt_dataengineers_materializations.snowflake_create_or_replace_table(target_relation, source_node)) %}
-        {% else %}
+        {% elif auto_maintained %}
             {% do build_plan.append(dbt_dataengineers_materializations.snowflake_create_or_replace_table(comparison_relation, source_node)) %}
         {% endif %}
-    {% else %}       
-        {% if current_relation is not none and  comparison_relation is not none %}
+    {% elif auto_maintained %}
+        {% if current_relation is not none and comparison_relation is not none %}
             {# If we are not doing a full refresh  #}
             {% if not full_refresh_mode %}
                 {%- set new_cols = adapter.get_missing_columns(comparison_relation, current_relation) %}
@@ -82,9 +83,9 @@
                         {% if new_col.name == old_col.name and new_col.data_type != old_col.data_type  %}
                             {% do build_plan.append(dbt_dataengineers_materializations.snowflake_alter_table_alter_column(current_relation, old_col.name, new_col.data_type)) %}
                         {% endif %}
-                    {% endfor %}    
+                    {% endfor %}
                 {% endfor %}
-                
+
             {% else %}
                 {% if source_node.external.migrate_data_over_flg %}
                     {% do build_plan.append(dbt_dataengineers_materializations.snowflake_migrate_data(migration_relation, target_relation, source_node)) %}
@@ -96,8 +97,6 @@
         {% do build_plan.append(dbt_dataengineers_materializations.snowflake_drop_table(comparison_relation)) %}
     {% endif %}
 
-    
-    
 
     {% do return(build_plan) %}
 {% endmacro %}