+++ macros/source_tables/snowflake/snowflake_create_schema.sql
@@ -1,0 +1,3 @@
+{% macro snowflake_create_schema(relation) %}
+    CREATE SCHEMA IF NOT EXISTS {{ relation.database }}.{{ relation.schema }}
+{% endmacro %}