+++ macros/streams/snowflake__stream.sql
@@ -9,15 +9,20 @@
   {%- set source_model = config.get('source_model') -%}
   {%- set source_schema = config.get('source_schema', default=schema) -%}
   {%- set source_database = config.get('source_database', default=database) -%}
+  {%- set source_database_prefix = config.get('source_database_prefix', default=none) -%}
   {%- set source_type = config.get('source_type', default='internal') -%}
-  
+
+  {% if source_database_prefix is not none %}
+    {% set source_database_var = source_database_prefix ~ "_" ~  target.name|replace('-', '_') %}
+    {% set source_database = var(source_database_var|replace("__", "_")|lower, database)   %}
+  {% endif %}
   {% set target_relation = this %}
   {% set source_relation = adapter.get_relation(identifier=source_model, schema=source_schema, database=source_database) %}
 
   {% if source_relation == none %}
-    {% set source_relation = api.Relation.create(identifier=source_model, schema=source_schema, database=database) %} 
+    {% set source_relation = api.Relation.create(identifier=source_model, schema=source_schema, database=source_database) %} 
   {% endif %}
-  
+
   -- setup
   {{ run_hooks(pre_hooks, inside_transaction=False) }}